requirements.txt
numpy==1.26.3
Pillow==10.2.0
streamlit==1.29.0
tensorflow==2.11.1
tensorflow_protobuf==2.7.0
